Enumerate method in UnmockRequest (#194)
* Move HTTPMethod to upper level interface.

* Change method in ISerializedRequest to be enum.

* Fix types in tests.

* Fix lint.
