+++ packages/unmock-core/src/service/service.ts
@@ -1,9 +1,9 @@
+import { isRESTMethod } from "../interfaces";
 import { DEFAULT_STATE_ENDPOINT, DEFAULT_STATE_HTTP_METHOD } from "./constants";
 import {
   ExtendedHTTPMethod,
   IService,
   IServiceCore,
-  isRESTMethod,
   isStateInputGenerator,
   StateInput,
   StateType,
